salary_data_pipelines/requirements.txt
apache-airflow==2.4.3
psycopg2-binary
pandas